update terminal msgs and remove and list fn
